secmet: move qualifier testfile into files per qualifier
